requirements.txt
python-dotenv
discord.py
langchain_huggingface
google.generativeai